+++ tests/utils/test_request_headers.py
@@ -1,3 +1,5 @@
+from http.cookies import SimpleCookie
+
 import pytest
 
 from alpha.utils.request_headers import Headers
@@ -17,6 +19,7 @@ from alpha.utils.request_headers import Headers
                 auth_token_type="Bearer",
                 refresh_token="refresh123",
                 api_key="apikey123",
+                _cookie_jar=None,
             ),
         ),
         (
@@ -28,6 +31,9 @@ from alpha.utils.request_headers import Headers
                 auth_token_type="Bearer",
                 refresh_token="cookie_refresh123",
                 api_key="cookie_apikey123",
+                _cookie_jar=SimpleCookie(
+                    "auth_token=cookie_auth123; refresh_token=cookie_refresh123; api_key=cookie_apikey123;"
+                ),
             ),
         ),
         (
@@ -40,6 +46,7 @@ from alpha.utils.request_headers import Headers
                 auth_token_type="Bearer",
                 refresh_token="cookie_refresh123",
                 api_key=None,
+                _cookie_jar=SimpleCookie("refresh_token=cookie_refresh123;"),
             ),
         ),
         (
@@ -52,6 +59,7 @@ from alpha.utils.request_headers import Headers
                 auth_token_type="Bearer",
                 refresh_token="refresh123",
                 api_key=None,
+                _cookie_jar=SimpleCookie("auth_token=cookie_auth123;"),
             ),
         ),
         (
@@ -64,6 +72,9 @@ from alpha.utils.request_headers import Headers
                 auth_token_type="Bearer",
                 refresh_token="cookie_refresh123",
                 api_key=None,
+                _cookie_jar=SimpleCookie(
+                    "auth_token=cookie_auth123; refresh_token=cookie_refresh123;"
+                ),
             ),
         ),
         (
@@ -73,6 +84,7 @@ from alpha.utils.request_headers import Headers
                 auth_token_type=None,
                 refresh_token=None,
                 api_key=None,
+                _cookie_jar=None,
             ),
         ),
     ],
@@ -102,11 +114,13 @@ def test_has_auth_token():
     assert headers.has_auth_token is True
     assert headers.auth_token_type == "Bearer"
     assert headers.auth_token == "abc.def.ghi"
+    assert headers.cookies == {}
 
     headers = Headers.from_headers({"Cookie": "auth_token=abc.def.ghi;"})
     assert headers.has_auth_token is True
     assert headers.auth_token_type == "Bearer"
     assert headers.auth_token == "abc.def.ghi"
+    assert headers.cookies == {"auth_token": "abc.def.ghi"}
 
 
 def test_has_refresh_token():